[BAU] refactor initial validation
